bkp/sic-001-mbpj-ghg-api/notifications/serializers.py
from datetime import datetime
from calendar import timegm
import json

from django.contrib.auth.forms import PasswordResetForm
from django.conf import settings
from django.utils.translation import gettext as _
from rest_framework import serializers
from drf_extra_fields.fields import Base64ImageField
from django.utils.timezone import now

from .models import (
    Notification
)

from users.serializers import (
    CustomUserSerializer
)


class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = '__all__'
        read_only_fields = ['id']

class NotificationExtendedSerializer(serializers.ModelSerializer):
    to_user = CustomUserSerializer(read_only=True)

    class Meta:
        model = Notification
        fields = '__all__'
        read_only_fields = ['id']