bkp/sic-001-mbpj-ghg-api/notifications/views.py
from django.shortcuts import render
from django.db.models import Q

from rest_framework.permissions import IsAuthenticated, AllowAny
from rest_framework.response import Response
from rest_framework.decorators import action
from rest_framework.filters import SearchFilter, OrderingFilter
from rest_framework import viewsets, status
from rest_framework_extensions.mixins import NestedViewSetMixin

from datetime import datetime

from django_filters.rest_framework import DjangoFilterBackend

from .models import (
    Notification
)

from .serializers import (
    NotificationSerializer, 
    NotificationExtendedSerializer
)

class NotificationViewSet(NestedViewSetMixin, viewsets.ModelViewSet):
    queryset = Notification.objects.all()
    serializer_class = NotificationSerializer
    filter_backends = (DjangoFilterBackend, SearchFilter, OrderingFilter)
    filterset_fields = ['title', 'message', 'to_user', 'date_sent']

    def get_permissions(self):
        if self.action == 'list':
            permission_classes = [AllowAny]
        else:
            permission_classes = [AllowAny]

        return [permission() for permission in permission_classes]    

    
    def get_queryset(self):
        queryset = Notification.objects.all()
        return queryset

    @action(methods=['GET'], detail=False)
    def extended(self, request, *args, **kwargs):

        queryset = Notification.objects.all()

        serializer_class = NotificationExtendedSerializer(
            queryset, many=True)

        return Response(serializer_class.data)
